refactor: Logo 동적 변경 기능 추가
기존 Header 컴포넌트에서 Logo부분을 분리한다.
각각의 서버마다 장바구니 정보가 다르므로 서버 이름이 장바구니 페이지 이름이 될 수 있도록한다.
